+++ libs/deal-ticket/src/deal-ticket-market.tsx
@@ -1,0 +1,59 @@
+import { addDecimal } from '@vegaprotocol/react-helpers';
+import { FormGroup, Input } from '@vegaprotocol/ui-toolkit';
+import { Market } from './deal-ticket';
+import { SideSelector } from './side-selector';
+import { SubmitButton } from './submit-button';
+import { TimeInForceSelector } from './time-in-force-selector';
+import { TypeSelector } from './type-selector';
+import { Order } from './use-order-state';
+import { VegaTxStatus } from './use-vega-transaction';
+
+interface DealTicketMarketProps {
+  order: Order;
+  updateOrder: (order: Partial<Order>) => void;
+  status: VegaTxStatus;
+  market: Market;
+}
+
+export const DealTicketMarket = ({
+  order,
+  updateOrder,
+  status,
+  market,
+}: DealTicketMarketProps) => {
+  return (
+    <>
+      <TypeSelector order={order} onSelect={(type) => updateOrder({ type })} />
+      <SideSelector order={order} onSelect={(side) => updateOrder({ side })} />
+      <div className="flex items-center gap-8">
+        <div className="flex-1">
+          <FormGroup label="Amount">
+            <Input
+              value={order.size}
+              onChange={(e) => updateOrder({ size: e.target.value })}
+              className="w-full"
+              type="number"
+              data-testid="order-size"
+            />
+          </FormGroup>
+        </div>
+        <div className="pt-4">@</div>
+        <div className="flex-1 pt-4" data-testid="last-price">
+          {market.depth.lastTrade ? (
+            <>
+              ~{addDecimal(market.depth.lastTrade.price, market.decimalPlaces)}{' '}
+              {market.tradableInstrument.instrument.product.quoteName}
+            </>
+          ) : (
+            '-'
+          )}
+        </div>
+      </div>
+      <TimeInForceSelector
+        order={order}
+        onSelect={(timeInForce) => updateOrder({ timeInForce })}
+      />
+      <SubmitButton status={status} market={market} order={order} />
+    </>
+  );
+};
