+++ libs/deal-ticket/src/deal-ticket.tsx
@@ -1,4 +1,4 @@
-import { FormGroup, Input, Dialog } from '@vegaprotocol/ui-toolkit';
+import { Dialog } from '@vegaprotocol/ui-toolkit';
 import { FormEvent, useEffect, useState } from 'react';
 import {
   Order,
@@ -7,15 +7,11 @@ import {
   OrderType,
   useOrderState,
 } from './use-order-state';
-import { ExpirySelector } from './expiry-selector';
 import { OrderDialog } from './order-dialog';
-import { SideSelector } from './side-selector';
-import { TimeInForceSelector } from './time-in-force-selector';
-import { TypeSelector } from './type-selector';
 import { useOrderSubmit } from './use-order-submit';
 import { VegaTxStatus } from './use-vega-transaction';
-import { addDecimal } from '@vegaprotocol/react-helpers';
-import { SubmitButton } from './submit-button';
+import { DealTicketMarket } from './deal-ticket-market';
+import { DealTicketLimit } from './deal-ticket-limit';
 
 const DEFAULT_ORDER: Order = {
   type: OrderType.Market,
@@ -141,120 +137,6 @@ export const DealTicket = ({
           id={id}
         />
       </Dialog>
-    </>
-  );
-};
-
-interface DealTicketMarketProps {
-  order: Order;
-  updateOrder: (order: Partial<Order>) => void;
-  status: VegaTxStatus;
-  market: Market;
-}
-
-const DealTicketMarket = ({
-  order,
-  updateOrder,
-  status,
-  market,
-}: DealTicketMarketProps) => {
-  return (
-    <>
-      <TypeSelector order={order} onSelect={(type) => updateOrder({ type })} />
-      <SideSelector order={order} onSelect={(side) => updateOrder({ side })} />
-      <div className="flex items-center gap-8">
-        <div className="flex-1">
-          <FormGroup label="Amount">
-            <Input
-              value={order.size}
-              onChange={(e) => updateOrder({ size: e.target.value })}
-              className="w-full"
-              type="number"
-              data-testid="order-size"
-            />
-          </FormGroup>
-        </div>
-        <div className="pt-4">@</div>
-        <div className="flex-1 pt-4" data-testid="last-price">
-          {market.depth.lastTrade ? (
-            <>
-              ~{addDecimal(market.depth.lastTrade.price, market.decimalPlaces)}{' '}
-              {market.tradableInstrument.instrument.product.quoteName}
-            </>
-          ) : (
-            '-'
-          )}
-        </div>
-      </div>
-      <TimeInForceSelector
-        order={order}
-        onSelect={(timeInForce) => updateOrder({ timeInForce })}
-      />
-      <SubmitButton status={status} market={market} order={order} />
-    </>
-  );
-};
-
-interface DealTicketLimitProps {
-  order: Order;
-  updateOrder: (order: Partial<Order>) => void;
-  status: VegaTxStatus;
-  market: Market;
-}
-
-const DealTicketLimit = ({
-  order,
-  updateOrder,
-  status,
-  market,
-}: DealTicketLimitProps) => {
-  return (
-    <>
-      <TypeSelector order={order} onSelect={(type) => updateOrder({ type })} />
-      <SideSelector order={order} onSelect={(side) => updateOrder({ side })} />
-      <div className="flex items-center gap-8">
-        <div className="flex-1">
-          <FormGroup label="Amount">
-            <Input
-              value={order.size}
-              onChange={(e) => updateOrder({ size: e.target.value })}
-              className="w-full"
-              type="number"
-              data-testid="order-size"
-            />
-          </FormGroup>
-        </div>
-        <div>@</div>
-        <div className="flex-1">
-          <FormGroup
-            label={`Price (${market.tradableInstrument.instrument.product.quoteName})`}
-            labelAlign="right"
-          >
-            <Input
-              value={order.price}
-              onChange={(e) => updateOrder({ price: e.target.value })}
-              className="w-full"
-              type="number"
-              data-testid="order-price"
-            />
-          </FormGroup>
-        </div>
-      </div>
-      <TimeInForceSelector
-        order={order}
-        onSelect={(timeInForce) => updateOrder({ timeInForce })}
-      />
-      {order.timeInForce === OrderTimeInForce.GTT && (
-        <ExpirySelector
-          order={order}
-          onSelect={(date) => {
-            if (date) {
-              updateOrder({ expiration: date });
-            }
-          }}
-        />
-      )}
-      <SubmitButton status={status} market={market} order={order} />
     </>
   );
 };
